# openapi.yaml
openapi: 3.1.0
info:
  title: Zantara Bridge API
  description: |
    Comprehensive API for Google Workspace integration, AI-powered chat, and productivity tools.
    The API provides secure endpoints for Drive, Gmail, Calendar operations, AI analytics, and conversation management.
    
    **Performance Features:**
    - Redis-based caching for Drive operations (5min folder cache, 10min file cache, 3min search cache)
    - Firestore-powered memory system with full-text search and relevance scoring
    - Automatic cache invalidation on data changes
    - Zero-cost optimized architecture with graceful fallbacks
  version: 1.0.0
  contact:
    name: Bali Zero HQ
    email: zero@balizero.com
  license:
    name: Proprietary
servers:
  - url: https://zantara-bridge-v2-prod-1064094238013.asia-southeast2.run.app
    description: Production server (Cloud Run)
  - url: https://zantara-light-bridge-1064094238013.asia-southeast2.run.app
    description: Zantara Light Bridge - Zero-cost optimized (Cloud Run)
  - url: http://localhost:8080
    description: Development server
security:
  - ApiKeyAuth: []
  - UserIdentification: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
      description: API key for authentication (required for most endpoints)
    UserIdentification:
      type: apiKey
      in: header
      name: X-BZ-USER
      description: User identification header (defaults to 'BOSS' if not provided)
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - Invalid or missing API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        ok:
          type: boolean
          default: false
        error:
          type: string
          description: Error message
        details:
          type: string
          description: Additional error details
      required:
        - error
...
    DriveFile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        mimeType:
          type: string
        webViewLink:
          type: string
        createdTime:
          type: string
          format: date-time
        parents:
          type: array
          items:
            type: string
...
paths:
...
  /api/search/recent:
    get:
      tags: [Search & Discovery]
      summary: Get recent files
      description: Retrieve recently modified files from Drive
      parameters:
        - name: limit
          in: query
          description: Maximum number of files to return
          schema:
            type: integer
            default: 5
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Recent files retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  recentFiles:
                    type: array
                    items:
                      $ref: '#/components/schemas/DriveFile'
                  count:
                    type: integer
  /actions/memory/save:
    post:
      tags: [Memory]
      summary: Save memory entry (Firestore Implementation)
      description: |
        Save a memory entry to Firestore with full metadata support.
        Supports tags, categories, relevance scoring, and automatic access tracking.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  description: User identifier for memory isolation
                  example: "user-123"
                content:
                  type: string
                  description: Main memory content
                  example: "Meeting notes about quarterly planning"
                title:
                  type: string
                  description: Optional memory title
                  example: "Q4 Planning Meeting"
                tags:
                  type: array
                  items:
                    type: string
                  description: Searchable tags
                  example: ["meeting", "planning", "q4"]
                category:
                  type: string
                  description: Memory category
                  example: "work"
                source:
                  type: string
                  description: Origin of the memory
                  example: "api"
                metadata:
                  type: object
                  properties:
                    tokenCount:
                      type: number
                    contextType:
                      type: string
                      enum: [conversation, document, note, task]
                  description: Additional metadata
              required:
                - userId
                - content
      responses:
        '200':
          description: Memory entry saved successfully to Firestore
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    default: true
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        description: Firestore document ID
                        example: "abc123def456"
                      message:
                        type: string
                        example: "Memory entry saved successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /actions/memory/search:
    get:
      tags: [Memory]
      summary: Search memory entries (Firestore Implementation)
      description: |
        Search stored memory entries in Firestore with full-text search, tag filtering, and time range queries.
        Supports relevance scoring and pagination for optimal performance.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
          description: User identifier for memory isolation
        - name: query
          in: query
          schema:
            type: string
          description: Full-text search query
        - name: tags
          in: query
          schema:
            type: array
            items:
              type: string
          description: Filter by tags
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
          description: Maximum number of results
        - name: minRelevanceScore
          in: query
          schema:
            type: number
            minimum: 0
            maximum: 1
          description: Minimum relevance score filter
        - name: timeStart
          in: query
          schema:
            type: integer
          description: Start timestamp for time range filter
        - name: timeEnd
          in: query
          schema:
            type: integer
          description: End timestamp for time range filter
      responses:
        '200':
          description: Memory search results from Firestore
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    default: true
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                            userId:
                              type: string
                            content:
                              type: string
                            title:
                              type: string
                            tags:
                              type: array
                              items:
                                type: string
                            category:
                              type: string
                            timestamp:
                              type: number
                            relevanceScore:
                              type: number
                            accessCount:
                              type: number
                        description: Array of matching memory entries
                      count:
                        type: integer
                        description: Number of results returned
  /actions/memory/stats:
    get:
      tags: [Memory]
      summary: Get memory statistics
      description: Retrieve comprehensive memory usage statistics for a user
      security:
        - ApiKeyAuth: []
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
          description: User identifier
      responses:
        '200':
          description: Memory statistics retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    default: true
                  data:
                    type: object
                    properties:
                      totalEntries:
                        type: integer
                      totalTokens:
                        type: integer
                      averageRelevanceScore:
                        type: number
                      oldestEntry:
                        type: string
                        format: date-time
                      newestEntry:
                        type: string
                        format: date-time
                      compressionRatio:
                        type: number
                      estimatedCost:
                        type: number
  /actions/memory/recent:
    get:
      tags: [Memory]
      summary: Get recent memory entries
      description: Retrieve the most recent memory entries for a user
      security:
        - ApiKeyAuth: []
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
          description: User identifier
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 50
          description: Maximum number of entries to return
      responses:
        '200':
          description: Recent memory entries retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    default: true
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                            content:
                              type: string
                            title:
                              type: string
                            timestamp:
                              type: number
                            tags:
                              type: array
                              items:
                                type: string
                      count:
                        type: integer
  /actions/memory/{id}:
    get:
      tags: [Memory]
      summary: Get memory entry by ID
      description: Retrieve a specific memory entry by its ID
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Memory entry ID
        - name: userId
          in: query
          schema:
            type: string
          description: User identifier for authorization
      responses:
        '200':
          description: Memory entry retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    default: true
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                      userId:
                        type: string
                      content:
                        type: string
                      title:
                        type: string
                      tags:
                        type: array
                        items:
                          type: string
                      category:
                        type: string
                      timestamp:
                        type: number
                      accessCount:
                        type: number
                      relevanceScore:
                        type: number
        '404':
          description: Memory entry not found
    
    put:
      tags: [Memory]
      summary: Update memory entry
      description: Update an existing memory entry
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Memory entry ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                title:
                  type: string
                tags:
                  type: array
                  items:
                    type: string
                category:
                  type: string
                relevanceScore:
                  type: number
                  minimum: 0
                  maximum: 1
      responses:
        '200':
          description: Memory entry updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    default: true
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                      message:
                        type: string
    
    delete:
      tags: [Memory]
      summary: Delete memory entry
      description: Delete a memory entry by ID
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Memory entry ID
        - name: userId
          in: query
          schema:
            type: string
          description: User identifier for authorization
      responses:
        '200':
          description: Memory entry deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    default: true
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                      message:
                        type: string
  /actions/memory/cleanup:
    post:
      tags: [Memory]
      summary: Cleanup old memory entries
      description: Remove old or low-relevance memory entries for cost optimization
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  description: User identifier
                maxEntries:
                  type: integer
                  description: Maximum entries to keep
                  default: 1000
                minRelevanceScore:
                  type: number
                  description: Minimum relevance score to keep
                  default: 0.1
                maxAgeInDays:
                  type: integer
                  description: Maximum age in days
                  default: 365
              required:
                - userId
      responses:
        '200':
          description: Cleanup completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    default: true
                  data:
                    type: object
                    properties:
                      deletedCount:
                        type: integer
                      message:
                        type: string
...